+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## vNext 
+
+- Add CAISO LMP Heat Map Example Notebook
+- Add a guide on contributing to gridstatus
+- Add Settlement Point Prices for ERCOT
+
 ## v0.16.0 - Dec 15, 2022
 
 - Ercot Fuel Mix Endpoint URL updated to include more fuel sources